update logic to only change debugger to true if local var wasn't set
