--- packages/design-system/src/components/CssDebugger/index.ts
@@ -1,1 +1,0 @@
-export { default as CssDebugger } from './CssDebugger';
