+++ packages/design-system/src/components/index.ts
@@ -4,7 +4,6 @@ export * from './Autocomplete';
 export * from './Badge';
 export * from './Button';
 export * from './ChoiceList';
-export * from './CssDebugger';
 export * from './DateField';
 export * from './Dialog';
 export * from './Drawer';
