+++ packages/design-system/src/styles/_set-for-deprecation.scss
@@ -1,11 +1,5 @@
 // Classes included in this file are set for deprecation and will be removed in the next major release.
 
-// The `--debugger` variable is used to show what soon-to-be deprecated classes are being used in the browser.
-// By default, the value is set to `false`. To enable the debugger, set the variable to `true` in the html selector.
-:where(html) {
-  --debugger: false;
-}
-
 // Include the soon-to-be deprecated classes here (remove them from their original files).
 // This makes it easier to find all the soon-to-be deprecated classes in one place and delete them.
 // For each class, include a comment above with the date the class was added to this file. This will help maintainers know how long a class has been listed as deprecated and when it should be removed from the project.
